Major update commit

+++ src/main/java/com/rifatsproject/Personalfinance/service/TransactionService.java
@@ -1,6 +1,8 @@
 package com.rifatsproject.Personalfinance.service;
 
+import com.rifatsproject.Personalfinance.domain.Account;
 import com.rifatsproject.Personalfinance.domain.Transactions;
+import com.rifatsproject.Personalfinance.exception.UserInputException;
 import com.rifatsproject.Personalfinance.repository.TransactionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,10 +15,26 @@ public class TransactionService {
 
     @Autowired
     private TransactionRepository transactionRepository;
+    @Autowired
+    private AccountService accountService;
 
 
-    public Transactions createTransaction(Transactions transaction){
-        return transactionRepository.save(transaction);
+    public Transactions createTransaction(Transactions transaction)
+            throws UserInputException {
+
+        String name = transaction.getAccountName();
+        Account account = accountService.getAccountByName(name);
+        if(account != null){
+            float currentbalance = account.getBalance() - transaction.getAmount();
+            if(currentbalance < 0){
+                throw new UserInputException("Operation unsuccessful");
+            }
+            account.setBalance(currentbalance);
+            accountService.updateAccount(account);
+            return transactionRepository.save(transaction);
+        }else {
+            throw new UserInputException("Operation unsuccessful");
+        }
     }
 
     public List<Transactions> getAllTransactions(){
@@ -38,10 +56,15 @@ public class TransactionService {
     }
 
     //delete
-    public String deleteTransaction(Long id){
-        String message = "Your transaction is deleted!";
-        transactionRepository.deleteById(id);
-        return message;
+    public String deleteTransaction(Long id) throws UserInputException{
+        Transactions t = transactionRepository.findById(id).get();
+        if(t.getTransactionName() != null){
+            String message = "Your transaction is deleted!";
+            transactionRepository.deleteById(id);
+            return message;
+        }else {
+            throw new UserInputException("Operation unsuccessful");
+        }
     }
 
     public float getMonthlyTransactions(){
